[DAOSM-39] vea: add additional unit and functional tests
Added unit tests to probe invalid parameters to the VEA API. Added asserts
in the VEA API to take a somewhat more defensive coding approach.

This commit requires a newer CART commit to take advantage of the ability
to replace assert() cmocka's mock_assert(). This enables us to test that a
function will throw an assert. Useful when testing invalid parameters.

Change-Id: Ib0614ab232dc9cdc48c424830d6233e87754a97f
Signed-off-by: Jeff Boyd <jeffrey.boyd@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33005
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>