DAOS-1233 vos: Fix vos unit test after recent punch rework
Conditional iteration isn't supported anymore when a range is
specified.  Disable the existing test for now so that we can get vos
tests to fail on future breakages.
Fix the vos_tests so that they return a non-zero exit code if a test
fails so it's caught when it starts failing rather than later.

Change-Id: I30a00a9209b6b290da56e93d376ae2f974f9d41a
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33019
Reviewed-by: Liang Zhen <liang.zhen@intel.com>
Tested-by: Liang Zhen <liang.zhen@intel.com>