DAOS-1229 evtree: Add evt_delete
Presently, it is simple.  It simply deletes the entry in the node
and updates the mbr accordingly.  If a node has no additional
entries, it is deleted.   This bubbles up to the root, deleting
the root if it is empty.
Adds adjust tree policy callback to move a rectangle when its
bounds are adjusted via insert or delete to maintain sorted order
in the tree.
User is responsible to free the associated data pointer that is
returned from evt_delete upon successful removal.
Update the debug code to indent each level in the tree

Change-Id: I421cc0ad520fa3769e91847fa642050b3d66bff7
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33109
Reviewed-by: Vishwanath Venkatesan <vishwanath.venkatesan@intel.com>
Reviewed-by: Jeffrey Boyd <jeffrey.boyd@intel.com>
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>