DAOS-1240 test: run_tests.sh swallows output on interruption/error
utils/run_test.sh is putting all of the test output as a test runs
into a file, and only after the test finishes, outputting the
contents of the file. The unfortunate side effect of this is that
no test output appears in the Jenkins console if the script is
interrupted before printing out the contents of that file (eg,
if it times out in Jenkins).

This was originally done as a way of filtering out any possible
"SUCCESS! NO FAILURES" messages that daos_post_build.sh is
looking for to determine test success in an individual test,
so it can instead create an aggregate success state at the end of
the script.

We can implement this in a way that "prints as it goes", so not
to lose output if the script is interrupted, which is important
in debugging, but still keep out the success message.

Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Change-Id: I3e418fa37ce67e15680d7accad703425d990fc8e
Reviewed-on: https://review.hpdd.intel.com/33037
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Phyllis Crandall <phyllis.crandall@intel.com>
Tested-by: Phyllis Crandall <phyllis.crandall@intel.com>