DAOS-1244 doc: Fix typo, usability issues in quickstart.md
The quickstart guide has a typo in it, where a { is instead a (
in a command.

While I am in the file, I am addressing some usability issues,
including making the fact there are some additional instructions
for developers more obvious, fixing text typos, and adding
--config=force to scons commands, along with a little
re-organizing to improve the order new people will encounter
information.

Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Change-Id: I101d442ff26e5a41b656d2780602de2e8e779413
Reviewed-on: https://review.hpdd.intel.com/33041
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-by: Stephen Willson <stephen.d.willson@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>