DAOS-1749 bio: support SPDK AIO bdev
Support SPDK AIO bdev. To use AIO bdev the env variable
VOS_BDEV_CLASS should be specified as 'AIO', and the AIO
bdev needs be configured in daos_nvme.conf.

Example of AIO device section in daos_nvme.conf

[AIO]
  AIO /mnt/daos/aiofile AIO1 4096
  AIO /dev/sdc AIO2

The first AIO bdev is constructed on a pre-allocated file, 4096
is the specified block size. the second AIO bdev is constructed
on a real device.

Signed-off-by: Niu Yawei <yawei.niu@intel.com>
Change-Id: I7bbcd993141c18e6c4fc1838eff1b5fb571046ea
Reviewed-on: https://review.hpdd.intel.com/33517
Tested-by: Jenkins
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>