DAOS-1238 test: Forgot to reactivate rdb tests in utils/run_test.sh in DAOS-1208
After changing rdb test runner code to align with output changes, I forgot to
uncomment the line in run_test.sh that actually ran the test.

In the course of getting this test to actually run during the Jenkins build an
issue arose when orterun could not find libnvme_discover.so. So the path was
added to LD_LIBRARY_PATH, LD_LIBRARY_PATH was added to the orterun command line,
and the test started passing again.

Change-Id: I069247cc30c2c8722b1569d2876e9171e413cc51
Signed-off-by: Phyllis Crandall <phyllis.crandall@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33047
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Reviewed-by: Omkar Kulkarni <omkar.kulkarni@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Li Wei <wei.g.li@intel.com>