DAOS-1791 vea: force free extents migration on ENOSPC
The extents freed by vea_free() won't be immediately visible
for allocation, they'll stay in a buffer for a short time and
being coalesced with other recent freed extents, then they'll
be batchly migrated back to compound index and visible for
allocation. The intention of this design is anti-fragmentation.

However, when vea_reserve() has already got ENOSPC, we'd force
a immediate free extent migration to try to satisfy the reserve
request.

This patch also fixed a typo in umem_init_txd().

Signed-off-by: Niu Yawei <yawei.niu@intel.com>
Change-Id: I6b8a6a4018e636863460d2933bd841fe8c3225cb
Reviewed-on: https://review.hpdd.intel.com/33562
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>