DAOS-1017 rdb: Clean up RDB Distributed Operations
RDB distributed start and stop operations call into the Pool service
which break modularity due to cyclical references (The Pool service is
a consumer of RDB). This patch moves the distributed start and stop
operations into the pool service layer.

Ideally, we'd like to have an RSVC (Replicated Service) library/module
from which the pool and container services are derived; at the moment,
both are hardwired to run as a single service. It is best to revisit
this issue when the two are separated in the future.

Signed-off-by: Omkar Kulkarni <omkar.kulkarni@intel.com>
Change-Id: Id3b663ed37f473183e56b3c40be7826263f15e8e
Reviewed-on: https://review.hpdd.intel.com/33082
Reviewed-by: Li Wei <wei.g.li@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>