CART-89 debug: Fix D_DEBUG to use temporary for d_log_check result
Rather than making a second call to d_log_check when we log a message
we can cache the result in a temporary variable.
This wasn't done originally because the macro is used everywhere and
can result in conflicts with user defined variables.
The compilers will detect this conflict if -Wshadow is enabled.
I added a pragma to enable this warning.  This method works for gcc
and clang but is ignored by icc.

CaRT is clean but this may require some changes to other projects.
I know daos has some shadowing, mostly of the form

int x;

if (y) {
	int x;
}

this isn't good practice anyway so it seems reasonable to force this
change on cart users.

Change-Id: I34c482bd83569a0ecb0daf4598c7c7caf57ca135
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33142
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins