DAOS-1304 evtree: Add evtree nodes to transactions
If a transaction aborts, we need to rollback evtree to old state.
This requires saving old state in pmdk.
This change uses the umem_tx api for transactions, avoiding the
longjmp and allowing transactional code to clean up state
Remove some unneeded tcx data structures

Adds return value for umem_free since pmemobj_tx_free returns
an int and there is often no need to continue if free failed.
This can happen in transactional free as it needs to add free
to the log.

Modifies pmem_tx_begin so that txd is optional.  It's not needed
for evtree transactions.

Change-Id: I9c1d11948078b7e5f970c4453f431499cad95eb3
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33156
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>