DAOS-1425 Pass environment to additional ranks in functional tests
Currently the functional tests assume that when starting servers on
multiple nodes, the environment of the user will be set sufficiently on
every node to allow the server to start and communicate with the rest of
the pool.

The environment from the initial server should be passed to the
additional servers through orterun's -x argument to alleviate this
assumption on remote node environments.

Take the opportunity to de-duplicate the LD_LIBRARY_PATH given to
orterun.

Also reduce the server start timeout to 30 seconds as experience is
showing that they take less than 3 seconds even on VMs.

Change-Id: Ib2a97188eb06fdf55fd975075a605b9e42d20caa
Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33181
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>