DAOS-1404 api: Add array size query to fetch API
1. Modify evt_get_max to evt_get_size.   Given that
offset is uint64_t, there is no way to return an
invalid offset other than to fail the routine.
Since the size is what we want and 0 is a valid size,
switch the API to return size instead.
2. Expose evt_get_size through fetch API using
iod_nr field.  Set it to 0 to get array size only.
Potentially, we can do the same for single value but
for now, we just handle the array size.
3. Modify the array addon to use this query rather
than enumerating rectangles from client.
4. Minor cleanup from prior reviews in evtree

Change-Id: I12fa59a0da48efe80b36fa96d02a1cf680fff387
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33204
Reviewed-by: Liang Zhen <liang.zhen@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins