DAOS-1311 build: Manage Go dependencies with dep
This patch uses the Go dep tool to manage Go dependencies.
All direct dependencies are listed in Gopkg.toml, pinned
to minor version or release branch in most cases. For forks
we control, we use master. For go.uuid, our code depends
on API changes after the last tagged release - for now we pin
to a specific commit, but we should change to a release tag
as soon as one is available.

The fetch_go_packages.sh shell script has been updated to
run "dep ensure" to make sure the vendor directory has
been populated with the appropriate versions.

Change-Id: I83cc8eb386141568b290858445b3e310b7672f8d
Signed-off-by: Kris Jacque <kristin.jacque@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33205
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>