DAOS-1355 bio: Update initial xstream count per device
This integrates the newly added capability to SMD to query the
device info to find the current number of xstreams mapped to the
given device. If the device is not currently in the device table,
then the count is set at 0. This inital xs count is important
in accurately assigning xstream->device mapping, which is determined
by the device with the least amount of mapped xstreams.

Change-Id: I3af4e71df9f826daf208a42d4e1496816f033d96
Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33187
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Reviewed-by: Vishwanath Venkatesan <vishwanath.venkatesan@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>