DAOS-1453 control: Use maps rather than slices in ControlService
In reference to code review comments on
https://review.hpdd.intel.com/#/c/33215/6, re-factor code to use maps
instead of slices for ControlService data in gRPC server. This change
makes sense as there are multiple methods that iterate through the
slices to find a given item and this would be simplified with maps.

Use this opportunity to increase code coverage and add relevant unit
tests.

  - Refactor gRPC server code to use maps rather than slices
  - Add test files for nvme and feature server module files
  - Expand test coverage for nvme and feature gRPC server logic

Change-Id: I9255adcf7e0e1d12157de98c52c3612fa0dc6f94
Signed-off-by: tanabarr <tom.nabarro@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33227
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Kris Jacque <kristin.jacque@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>