DAOS-1435 build: Add fio dependency to build for spdk
Control plane feature to run burn-in on selected NVMe SSD
controllers is to be enabled through SPDK fio plug-in. fio
built binaries need to be made available as well as src
being made available when building spdk.

This patch just enables fio to be included in scons build
so spdk fio_plugin will be built and made available for
use by control plane.

  - Update scons_local commit to build fio as prereq of spdk
  - Provide SHA1 for FIO in util/build.config
  - Update quickstart.md with details of fio dependency

Change-Id: I55df19e208360a6fe97d6ac9a3a75a2a57b882fb
Signed-off-by: tanabarr <tom.nabarro@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33250
Tested-by: Maloo <hpdd-maloo@intel.com>
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>