DAOS-623 build: Remove hard coded references to variant dir
Using 'build/' is not reliable in SConsripts because it
assumes the variant dir is fixed.   One change to top level
SConscript would break this assumption.   So, this modifies
those references to use paths based on current directory.
Similarly, it removes hard coded paths using '#' preferring
relative paths instead.   This is less error prone and more
robust to directory changes.

This can be cleaned up further by switching internal headers
to "header.h" style instead of <header.h> but that is a bigger
change.

Change-Id: Ie2e5f79725a77676814c4b2905b86900defcaae3
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33286
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>