CART-531 locking: Fix incorrect lock handling in list iteration.
It's not safe to drop the lock during d_list_for_each() loops,
so do not drop the lock for the code blocks.

This may deadlock, as I haven't checked the registered callbacks
yet for any list manipulation.

Switch to using the non _safe versions of this calls as there's
no indication that the lists are being modified during the calls.

Change-Id: I60345c4eb63c1429182fac2e2fdde8f38cdf4c6d
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33304
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>