DAOS-1543 Add support for program-only dependencies
Currently in our scons configuration, we have no way to specify how
a component that only provides binaries (i.e. programs) can be
discovered in the same way library type components do with libs and
headers.

Add a progs option to _Component to specify programs to look for to
determine if a component is installed already.

Change-Id: If298b6c1fd80750adce490ef9fd7cdeca92a711e
Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33321
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>