DAOS-1255 bugfix: Change output when server failure to start
When the dss server fails to start there is no output in
the log to indicate there may be a problem relating to
the file system type or whether or not it is mounted.

A warning was added to the log when the dss_svr_init call
fails in order to remind the user to check that the correct
file system type is mounted at the provided location.

Change-Id: I621edf6035869cf39a3e6ce4f47b5e8e263f17ab
Signed-off-by: Peter A. Fetros <peter.fetros@intel.com>

Added a missing tab.

Change-Id: I621edf6035869cf39a3e6ce4f47b5e8e263f17ab
Signed-off-by: Peter A. Fetros <peter.fetros@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33405
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>