DAOS-1973 Add the abilty to add specific tests to avocado
In addition to the tests found by tag, we will want a way to
specify specific tests regardless of their tag(s).

This will be useful for example when a patch adds or modifies
a test that is not part of the "regression" suite so as to force
it to be run during the patch's CI test run.

Change-Id: Ibdfbf0c948c0942673599ed74147509ae96ce3f7
Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33834
Tested-by: Jenkins
Reviewed-by: Saurabh Tandan <saurabh.tandan@intel.com>
Reviewed-by: Scott Kirvan <scott.kirvan@intel.com>