CORCI-379 build: Add ProgramBinary to define aliases for required binaries
A while ago a regression was added to components.py in the sense that
the build of any component would fail if ninja wasn't installed on the system
even if the build did not require ninja.   This would be an issue for
components like cart that don't use FUSE.   This patch adds a new class
to help with this issue.

Instead of checking upon executing components.py, the check is now deferred
until the required_progs configure check is done and $name is set
in the construction environment so the component definition can use it.

It's possible to generalize this for libraries or other scenarios but this
was a low hanging fruit to fix the immediate issue.

Change-Id: I17ba6427f3ae4ae4d7159d155a878544308a6751
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33473
Tested-by: Jenkins
Reviewed-by: John Malmberg <john.e.malmberg@intel.com>