CART-89 htable: use htable_create_inplace for group code.
The group membership code creates a large number of hash tables
at startup, so rather than allocate memory seperately, and keeping
a pointer array for each one, just create an array of hash tables
and create them inplace.

This should be slightly quicker as it's one less level of
indirection, however it's also faster to startup and test,
particuarly when doing the fault injection tests which loop
for every memory allocation during startup.

Change-Id: I47145ae8a8c34833b6927b8b7a7c2ba9c52b177d
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33477
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>