DAOS-1636 vos: Keep update range for dkey/akey
Storing the punched epoch only is insufficient for detecting
errors in vos updates.   If we store the range instead, we
can detect some issues.

This isn't a full implementation.   It should essentially
retain the existing behavior except that it would disallow
non-rebuild punch to earlier epoch than the latest update.
As it stands, it only does this if the key was previously
punched.

Change-Id: I6a5fb015fdf4fcae31a3d867c7cf5f94b474bce7
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33479
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>