DAOS-1636 vos: Keep update range for an object
This is a followup to https://review.hpdd.intel.com/#/c/33479/

It does the same thing for the object tree, keeping track of
the range of updates.  It doesn't change any behavior for
iteration yet but it is a precursor to being able to do
conditional iteration of keys.

Modifies the timestamp update code so it can use same code
for object and key timestamp fields.

Change-Id: I9dd83b59eb8940e3cdbf2eed84a95a6aabc713d2
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33494
Tested-by: Jenkins
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>