DAOS-1724 server: Perform MS replica check
During the system formatting/provisioning phase, a server needs to check
if it is supposed to host an MS replica, by comparing each access point
to the address it listens on. This patch implements such a check but
only prints the result, leaving the actual formatting of the MS to
follow-on patches.

We currently only support IPv4 addresses.

Change-Id: I7fe2efeaf746b09f403fe9338438667a05fee6fb
Signed-off-by: Li Wei <wei.g.li@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33503
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kris Jacque <kristin.jacque@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>