DAOS-1636 vos: Optimized conditional iteration for objects
Now that we store the range of known updates, we can optimize
conditional iteration to skip objects that are not visible
in the specified range, particularly if created after the range.

Having the earliest known update allows us to avoid some more
complicated logic to return only the first incarnation punched
after the range.

Changes boundary condition for low end of range to match that
of dkey/akey iteration.   If an object is punched in epc_lo,
it will not be returned.

Change-Id: Ied7cb09b93130652846c817e2cb428eaf00b0a13
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33523
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>