DAOS-1427 common: Add C support for drpc listener
This patch adds an API to set up a socket to listen
for dRPC messages. The listening socket is set up
as non-blocking.

The progress method blocks with a timeout, and
maintains the connections and listener. It accepts
new connections on the listener, and processes
messages for connected sessions. If a session fails
or is closed by the peer, it is cleaned up. If a
listener receives any errors, it's a surprise --
we leave the sessions open and propagate up an
error code for the caller to address.

A simple example flow:

1. Create the listening socket: drpc_listen()
2. Create a drpc_progress_context containing the
   the listening socket's drpc context.
3. Call drpc_progress() with a timeout. This
   calls drpc_accept() and/or drpc_recv() as
   needed if incoming activity is detected on
   the open sockets.
4. Check if a shutdown condition is detected.
5. If a shutdown condition was detected,
   drpc_close() all open sockets and exit.
   Otherwise go back to 3.

Change-Id: Id6b9cc2636dd8d5da8a02cec88da4c69a7e72013
Signed-off-by: Kris Jacque <kristin.jacque@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33721
Tested-by: Jenkins
Reviewed-by: David Quigley <david.quigley@intel.com>
Reviewed-by: Li Wei <wei.g.li@intel.com>
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>