DAOS-1682 vos: Add support for nested iterator
This is a simplification of vos iteration as it enables
initializing an iterator from its parent iterator.

For example, if you are iterating dkeys, you can prepare
an iterator on the akey tree at the cursor using the
cursor rather than running tree_prepare.   This avoids some
of the problems of key_tree_prepare such as ensuring you pass
the correct epoch to key_tree_prepare to avoid missing
a punched key that already matched the condition.

Currently only supports OBJ->DKEY->AKEY->RECX|SINGLE

Also modified evt_iter_prepare to take a range and
evt_move_trace to also take an optional range so
we can skip rectangles that aren't part of our
selection.  This was part of my debugging of this
update but turned out to not be the issue.  I can
separate it if preferred.  It's a precursor to
better evt_find.

Updated daos_perf shows no measurable difference

Needs more testing for VOS_ITER_OBJ

Change-Id: I6f072b5684506a94cf0022949868ffe51ec86ee0
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33592
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>