DAOS-1862 evtree: Refactor evt_find logic
Ultimately, we want to be able to iterate through rectangles
in evtree.   This is a step in that direction by simplifying
the logic dealing with evt_ent_list.   Instead of allocating
chunks as needed, if we go over the embedded limit, we
just allocate enough space to fit all of the rectangles
including any splits of a rectangle.   This is 2x the size of
the evtree as inserting one rectangle can at most increase
the visible rectangle count by 1.

This change simplifies a lot of logic and prepares for
implementing evt_find functionality in the generic
iterator.

Change-Id: Iea77960c5e1d1ea7ca402f6bda661c4cd5fe634b
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33671
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>