CART-89 proc: Make some hg_proc functions static, and remove proc_memcpy on zero length arrays
Identify a few cart functions that are defined in internal header files
but only ever called from the same file, so remove the prototypes and
make the functions static.

Remove a call to hg_proc_memcpy() on zero length array as it doesn't
seem to make sense and will add to the packet size, the only possible
thing it could do would be to transmit a value using ca_arrays if
ca_count was non-zero however this seems a horrible API misuse so do
not allow it.

Change-Id: Ied5d5043adf21177938519d56057b5f5ddc4ffa0
Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33731
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Tested-by: Jenkins
Reviewed-by: Alexander Oganezov <alexander.a.oganezov@intel.com>