DAOS-1423 security: Add infrastructure to validate security credentials
This patch moves the pool attribute structure into a header file so it
is accessible for the access control checks in the security module. It
also includes the security api call for the pool service to use during
an access check. It does not however change the pool connect to use
the new credential at this point.

Change-Id: Ic1ff1464afdfdbeb7b2b221639f417e6a2ad6bd3
Signed-off-by: David Quigley <david.quigley@intel.com>