DAOS-1862 evtree: Fix a bug in evtree iteration/find
In reorganizing the code, the memcpy got removed in array
reallocation just resulting in random garbage.  It appears
our testing doesn't tickle this case (something that can be
fixed with an evtree test driver).  It does actually test it
but D_ALLOC_ARRAY sets the memory to 0 so the first entries
after an allocation will be empty extents.  Since we only
presently test iteration visually and with lack of segv, it
was lost in noise.

It did show up when I wrote this patch
https://review.hpdd.intel.com/#/c/33797 and tried to iterate
over 1000 objects where we count the number of records found.

Change-Id: I195cf5e3c8f5ced63dcefdb73ec3266f15862c47
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33826
Tested-by: Jenkins
Reviewed-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>