DAOS-1995 container: Remove punch of uncommitted data on cont_close.
On Container close, we currently punch anything that is not committed.
this does not work well with the new epoch model changes, since the
application does not manage epochs anymore. remove this punch
making everything visible after container close, even uncommitted
transactions.

When two phase commit is available, things will be as expected,
with uncommitted txs being non visible, but for now this is a better transition.

Also remove the temporary daos_cont_sync() since it's not required anymore.

Change-Id: I81509295970fc32d3bf5a257254fae40408481f6
Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33874
Tested-by: Jenkins
Reviewed-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>