CORCI-413 Jenkinsfile and support for DAOS
Between the need to move to GitHub from Gerrit, and losing our Autotest
and Maloo developers in the sale of Lustre, we find ourselves needing to
re-do most of our CI pipeline.  It does not make sense to continue to
use Autotest and Maloo without support as they are maintenance heavy
systems designed for a different use case than our own.

To replace it, we are building a new CI pipeline that will use GitHub as
the main repository,and then provisioning and testing will be triggered
from Jenkins directly. We will make use of the Jenkins pipelines feature
to more easily facilitate this model in Jenkins.

Change-Id: I4462957fc3a3f1ef1a7925645ec03f8de6224be0
Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>