DAOS-2027 vos: Avoid adjusting the epoch range for array iterator
The epoch range was previously only valid for single value.  This
change restores that behavior.
The user is responsible for getting the correct epoch range to
iterate what they want for recx.
The only leftover functionality that we need is reverse iteration
for arrays and that can be handled by adding an additional
flag for the recx condition.


Change-Id: I4a7b6e0b2b2bea0f57306877a910c5dee5d41e5c
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33935
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>