DAOS-2050 evtree: Implement evt_iter_delete/empty
For delete, it only is supported for non-sorted iterator.
The function will keep trace intact though reprobe would
be needed if the user yields between calls deleting the
returned payload, for example.  But by default, the trace
will be set to the next entry.

Add some unit test infrastructure and a test for this
feature.  Will continue to add some additional tests.

Bug fix
1. The filter can't filter must not filter intermediate
   nodes on the low epoch because the in-tree rectangle
   only contains the lowest epoch of all rectangles.
2. Call evt_desc_free to free payload if user passes
   NULL to evt_delete or evt_iter_delete
3. Fix a few persistance issues caught by valgrind

Change-Id: Iff45b596e2649eabf4e86a5b2aad9962679d9e07
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33952
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Tested-by: Jenkins
Reviewed-by: Ryon Jensen <ryon.jensen@intel.com>