DAOS-1967 init: call crt_init_opt() with correct sep parameters
When using SEP (scalable endpoint), if user set un-reasonable
CRT_CTX_NUM it possibly cause failure when creating cart context.
This patch refine it by calling crt_init_opt() with corrected sep
parameter when needed.

With this patch, the ENV of CRT_CTX_NUM is obsolete. Another change
in this patch is, sockets provider will always use regular EP rather
than SEP because there is no benifit and it has hard limit on SEP
which is exposed with the threading model change.

Change-Id: If68d802401aa61b604c9d3b2e93a11a2755ea852
Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33946
Tested-by: Jenkins
Reviewed-by: Liang Zhen <liang.zhen@intel.com>