DAOS-2035 vos: Expose original extent and visibility flags in VOS
The extents returned by the vos iterator aren't very useful without
this information.   We need to know the bounds of the original
in-tree extent to do delete and we need  to know if the extent
is visible or not.  Also add EVT_LAST and expose it to mark final
entry in sorted array.

Change-Id: Iab65eae564f9c9b38a65d788449281f518374df8
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33948
Tested-by: Jenkins
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Reviewed-by: Ryon Jensen <ryon.jensen@intel.com>