DAOS-2044 control: merge mgmt and server packages
Re-factor to merge mgmt package into server, seems to be preference
in Go for flatter package structure. Server package is the only
consumer of mgmt so makes sense to merge.

Move proto to top-level dir so it can be consumed by multiple
packages (client & server) and cliOptions struct back to
daos_server.go.

Change-Id: I7d4210f94c34b2b2bcf193b2bfd8114526c046bc
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: https://review.hpdd.intel.com/33981
Tested-by: Jenkins
Reviewed-by: Li Wei <wei.g.li@intel.com>
Reviewed-by: David Quigley <david.quigley@intel.com>
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>