DAOS-1441 dtx: DAOS server two-phase commit transaction (DTX)
The main purpose of this project is to guarantee the consistency
among replicas when update an DAOS object with multiple replicas.

The basic idea is as following:

When client wants to modify an object with multiple replicas the
client sends the modification RPC to the leader repica (selected
via some algorithm). Then the leader dispatches the RPC to other
non-leader replicas. And then all the replicas can make each own
modification locally in parallel.

For each replica, before the local modification, it will start a
server-level transaction (called as 'DTX') firstly, then related
modifications will be attached to such DTX. For non-leader shard,
after the local modification done, it will mark its DTX state as
'prepared', then reply to the leader replica. For leader replica,
if all replicas have done the modification, it will mark the DTX
state as 'committable' (locally), and then reply the client with
modification done. For the client, as long as it receives leader
reply with modificaction done, then it will regard it as succeed
in spite of whether the modification is really committed or not.
It is the leader's duty to committed the 'committable' DTXs some
time later or if 'committable' DTXs count exceeds some threshold.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I226a20b264936abda82dffe5752f9238945d7dfb