DAOS-2070 vos: Fix some issues with PMDK
There are a few bugs with our usage of PMDK and still some more
to fix.   This change fixes a few in btree and evtree, primarily
with in-transaction stores without adding memory to a transaction.

The other primary change here is to create a simple helper
library to be able to test with pmdk or virtual memory.  It would
probably be good to add nvme to the mix here but I'll defer that
to later.   As I was trying to find some issues with btree, I
noticed that inplace updates didn't work well with PMDK and I had
recently added similar changes to evt_ctl so I thought I'd
commonize those changes.

I plan to also add some more normal unit tests and this
infrastructure should help in that regard.

Add code to shell scripts to run with special valgrind tool. This
doesn't work with normal valgrind as pmemcheck is unavailable as
of yet.

Change-Id: I23f4946a321ee99aafa6f3758c7c23bb727eb1f6
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>
Reviewed-on: https://review.hpdd.intel.com/34042
Reviewed-by: Ryon Jensen <ryon.jensen@intel.com>
Tested-by: Jenkins
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>