DAOS-1441 vos: DTX Commit-on-Share cache in VOS
For performance consideration, we commit the DTX asychronously.
So for most cases, when the DTX become committable, it will be
cached on the leader. The leader will commit multiple DTXs via
single DTX commit RPC under the following cases:

1. The cached committable DTXs count exceeds some threshold (512).

2. Some cached committable DTXs become too old (60 seconds).

Only above two conditions for aggregated DTXs commit is not enough.
Because the DTX state affects related data record visibility. It is
possible that before the committable DTX is committed, some new RPC
comes which target is the same or related with the data record that
has been modified via some cached committable DTX. In the DTX model,
only the leader know whether the DTX is committable or not. Then we
need to commit related (called as 'share' here) committable DTXs in
advance before handling the new modification.

The CoS cache is used for the detection of DTX sharing or conflict.
All the committable DTXs are cached on the leader in DRAM with new
B+ tree that is attached to the container. The key for the tree is
consisted of the object ID plus dkey hash. So it is per-dkey based
cache. Each CoS record has two local lists: update list and punch
list, for update DTXs and punch DTXs separated. The DTXs that modify
different akeys (under the same dkey) are linked into the local list
(either punch list or update list).

To simplify the logic, if the modification only tocuh the object,
such as DAOS_OBJ_RPC_PUNCH, the DTX will be committed immediately
when it becomes committable.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Ibd656bcc20bd03ee3d295174e82ea25cb309481e